refactor(I18n): Extract translate and createUseI18n in file
no code change
