+++ react/I18n/index.jsx
@@ -4,9 +4,8 @@
 
 'use strict'
 
-import React, { Component, useContext, useMemo, forwardRef } from 'react'
+import React, { Component, useContext } from 'react'
 import PropTypes from 'prop-types'
-import Polyglot from 'node-polyglot'
 
 import { initTranslation } from './translation'
 import { initFormat } from './format'
@@ -76,53 +75,12 @@ I18n.childContextTypes = {
   lang: PropTypes.string
 }
 
-// higher order decorator for components that need `t` and/or `f`
-export const translate = () => WrappedComponent => {
-  const Wrapper = forwardRef((props, ref) => {
-    const i18nContext = useContext(I18nContext)
-
-    return (
-      <WrappedComponent
-        {...props}
-        ref={ref}
-        t={i18nContext && i18nContext.t}
-        f={i18nContext && i18nContext.f}
-        lang={i18nContext && i18nContext.lang}
-      />
-    )
-  })
-
-  Wrapper.displayName = `withI18n(${WrappedComponent.displayName ||
-    WrappedComponent.name})`
-
-  return Wrapper
-}
-
 export const useI18n = () => {
   return useContext(I18nContext)
 }
 
-export const createUseI18n = locales => () => {
-  const { lang } = useI18n() || { lang: DEFAULT_LANG }
-  return useMemo(() => {
-    const polyglot = new Polyglot({
-      locale: DEFAULT_LANG,
-      phrases: locales[DEFAULT_LANG]
-    })
-    if (lang && lang !== DEFAULT_LANG) {
-      try {
-        polyglot.locale(lang)
-        polyglot.extend(locales[lang])
-      } catch (e) {
-        console.warn(`The dict phrases for "${lang}" can't be loaded`)
-      }
-    }
-    const f = initFormat(lang)
-    const t = polyglot.t.bind(polyglot)
-    return { t, f, lang }
-  }, [lang])
-}
-
 export { initTranslation, extend } from './translation'
+export { default as translate } from './translate'
+export { default as createUseI18n } from './createUseI18n'
 
 export default I18n
